BUG: vtkMRMLLinearTransformNode::SetAndObserveMatrixTransformToParent was not firing vtkMRMLTransformableNode::TransformModified
Simplify SetAndObserveMatrixTransformToParent to make sure vtkCommand::ModifiedEvent and vtkMRMLTransformableNode::TransformModified are fired.
Ensuring that vtkMRMLTransformableNode::TransformModified is called allows vtkMRMLTransform node observers to be notified when a new vtkMatrix4x4 is set and when the matrix is changed. vtkCommand::ModifiedEvent is fired for all vtkMRMLNode changes (including when a new Matrix is set) but not when the matrix is modified.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19927 3bd1e089-480b-0410-8dfb-8563597acbee