ENH: fix nrrd/vtkITK reader tests for tensors, scalars
One remaining issue is that tensor measurement frame
is being read by ITK as LPS but needs to be RAS for slicer
purposes, so the first two columns need to be negated.

Note that since these two tests are not in the form of a
module, there is no need to install them in the binary
tree, which is why the cmake lists was changed.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20772 3bd1e089-480b-0410-8dfb-8563597acbee