BUG: #2418 - manage python and C++ access to the dicom database
It might be possible to remove slicer.dicomDatabase and change
all code to use slicer.app.dicomDatabase() as done in the branch
linked below.  However it's not clear how to manage reference counts.
When a ctkDICOMDatabase is instanced in python and then set in the
core application with setDICOMDatabase things are fine.  But calling
setDICOMDatabase(None) destroys the ctkDICOMDatabase even though
the PythonQt wrapper still exists.  Calling methods on the PythonQt
wrapper gracefully refuses to operate on a destroyed instance.  Until
this issue is cleared up, it's better for python to retain an instance
of the database and pass it down to the core application only when it
is known to be valid.

https://github.com/pieper/Slicer/tree/2418-dicomdb-in-core-app-remove-slicer.dicomDatabase

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20832 3bd1e089-480b-0410-8dfb-8563597acbee