ENH: fix implementation of Slicer Data Bundle writing
Rather than make a new temp scene, we update the paths
of the existing scene and restore them afterwards.  Creating
the temp scene caused problems with IDs not matching references
and this was not correctable since nodes were shared between
scene views and the original nodes (that is, remapping all the
nodes inside scene views to have the nodes they needed to
have for the databundle save caused problems for the original scene).

Fixes #2120

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20903 3bd1e089-480b-0410-8dfb-8563597acbee