ENH: Add qMRMLTreeView::showScene and qMRMLTreeView::showRootNode
Give more flexibility to decide what item is root.
A root index for QTreeView is the invisible index that is the parent of all
the top level indexes.
showScene or showRootNode give the option to decide if the root scene/node
is visible, which implies that its parent becomes the root index for QTreeView.