ENH: Improve transform interaction
Now the transform will automatically be centered around whatever objects
it transforms at the moment. This is more intuitive to the user as
it looks as if the object under the transform itself is
moved/scaled/translated.
When no object is present, the transform is just scaled to facilitate
interaction with it.

Testing is added to ensure that:
 - The widget holds correctly the objects it will transform. Furthermore this
test is perfomed with models to extend the coverage as transform interaction
can theoritically deal with any vtkMRMLDisplayableNode.
 - Testing for interacting with transforms with parents
 - Serialization and deserialization of the scene