ENH: Remove QWebChannel warnings
The QWebChannel emits a lot of warning in the form of:
" Property OBJ::METHOD() of object OBJ has no notify signal and is not constant, value updates in HTML will be broken!"
This is due to registering two object to javascript:

 - qSlicerExtensionsManagerModel:
We can solve this issue by cleaning up the properties on the object and
making sure all properties have the proper NOTIFY signal.
Properties where it would not be appropiate to have a signal have been
removed. In those cases, the method were made invokable so python could
still access them (where needed).

 - qSlicerExtensionInstallWidget:
Because we cannot control all the properties on the object (as most of them
are inherited), we create a pass through class that is used to pass along
each of the method form javascript to C++